+++ pkg/apis/kubernetes/v1/xxgenerated_object.go
@@ -11,6 +11,11 @@ func (o *Cluster) GetIdentifier(ctx context.Context) (string, error) {
 	return o.Identifier, nil
 }
 
+// GetIdentifier returns the primary identifier of a kubeconfig object
+func (o *kubeconfig) GetIdentifier(ctx context.Context) (string, error) {
+	return o.Cluster, nil
+}
+
 // GetIdentifier returns the primary identifier of a NodePool object
 func (o *NodePool) GetIdentifier(ctx context.Context) (string, error) {
 	return o.Identifier, nil